Approximate charges and its positions
Given a data of measured electric potential of four (2 opposite positive charged balls), and (2 opposite negative charged balls), and the data set containing x,y points for insulated tray where the experiment was performed. The task of this project was to use curve_fit() model to approximate the charges and positions of those charged balls on the tray. Heat-map was use to visualize x,y and potential data sets.